# definition/api.yaml
openapi: 3.0.3
info:
  title: PramTube API
  description: Self-hosted video streaming platform API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
    description: Local development
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: philos_session
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
...
    ResponseComment:
      type: object
      required:
        - id
        - video_id
        - user_id
        - user_name
        - body
        - created_at
      properties:
        id:
          type: string
        video_id:
          type: string
        user_id:
          type: string
        user_name:
          type: string
        user_picture:
          type: string
        body:
          type: string
        created_at:
          type: string
          format: date-time
...
    VideoStatusEnum:
      type: string
      enum:
        - uploading
        - processing
        - ready
        - failed
    ResponseVideo:
      type: object
      required:
        - id
        - title
        - visibility
        - status
        - play_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
        user_id:
          type: string
        uploader_name:
          type: string
        uploader_picture:
          type: string
        title:
          type: string
        visibility:
          type: string
          enum:
            - private
            - unlisted
            - public
        status:
          $ref: '#/components/schemas/VideoStatusEnum'
        width:
          type: integer
        height:
          type: integer
        duration:
          type: string
        codec:
          type: string
        hls_path:
          type: string
        size_bytes:
          type: integer
          format: int64
        thumbnail_path:
          type: string
        play_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  parameters:
...
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
paths:
...
  '/api/v1/videos/{id}':
    get:
      operationId: GetVideo
      summary: Get a video by ID
      tags:
        - videos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseVideo'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: DeleteVideo
      summary: Delete a video
      tags:
        - videos
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Video deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Video not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: UpdateVideo
      summary: Update video properties
      tags:
        - videos
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                visibility:
                  type: string
                  enum:
                    - private
                    - unlisted
                    - public
      responses:
        '204':
          description: Video updated
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/api/v1/uploads/{upload_id}/thumbnail':
    post:
      operationId: UploadThumbnail
      summary: Upload a thumbnail image for a video
      tags:
        - uploads
      security:
        - CookieAuth: []
      parameters:
        - name: upload_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - thumbnail
              properties:
                thumbnail:
                  type: string
                  format: binary
      responses:
        '204':
          description: Thumbnail uploaded
        '400':
          description: Invalid image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  '/api/v1/videos/{video_id}/comments':
    get:
      operationId: ListComments
      summary: List comments for a video
      tags:
        - comments
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseComment'
    post:
      operationId: AddComment
      summary: Add a comment to a video
      tags:
        - comments
      security:
        - CookieAuth: []
      parameters:
        - name: video_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - body
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '201':
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseComment'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...